# Conflicts:
#	ERD/scsSql.erm
#	ERD/scsSql_data.sql
#	src/main/java/com/lec/spring/controller/BoardController.java
#	src/main/java/com/lec/spring/domain/Post.java
#	src/main/java/com/lec/spring/domain/User.java
#	src/main/java/com/lec/spring/repository/PostRepository.java
#	src/main/java/com/lec/spring/service/BoardServiceImpl.java
#	src/main/java/com/lec/spring/service/CategoryServiceImpl.java
#	src/main/java/com/lec/spring/util/U.java
#	src/main/resources/mapper/PostRepository.xml

+++ src/main/java/com/lec/spring/domain/Authority.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class Authority {
-    private int id;    // PK
+    private Long id;    // PK
     private String name;    // 권한명 (ADMIN, MEMBER)
 }
